app/pyproject.toml
[project]
name = "ss2gdrive"
version = "0.3.4"
description = "Simple Shot to Google Drive (Wayland/Flatpak) with Tray & Settings"
requires-python = ">=3.10"
dependencies = [
  "google-api-python-client>=2.149.0",
  "google-auth>=2.35.0",
  "google-auth-oauthlib>=1.2.1",
  "dbus-next>=0.2.3",
  "PySide6>=6.7.2",
]

[project.scripts]
ss2gd = "ss2gd.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
